handle non-boolean case clauses in switch statement with empty tag
Fixes #204

For switch statements with no tag, we effectively just cast each case
expression to a bool. But a case expression can have type interface{},
and casting the javascript equivalent to a bool always returns true. To
fix that, this commit converts such an expression X to the expression
"true == x" before translation to javascript.