fix for goroutine blocking bookkeeping
I imagine this problem shows up elsewhere, but the clearest
demonstration I can find is in the behavior of runtime.NumGoroutine():
The second time a goroutine blocks, and every time thereafter, it
invokes the goroutine exiting logic. This causes runtime.NumGoroutine to
underreport the number of goroutines, because the exiting logic includes
decrementing that count. In my testing, runtime.NumGoroutine was often
negative.

The fix is pretty simple. When calling the `r.$blk` closure, return its
output, which may itself be another closure.