gopherjs serve: Display compilation errors in browser console.
This is nice when using with gopherjs serve, because it already
recompiles on browser refresh. This allows one to switch between code
editor and browser for a faster iteration cycle. No need to also switch
to the terminal to see errors, as now both compilation and runtime
errors are visible in browser console.