gopherjs serve: Fix precedence of served files.
Previously, if there was already a {base}.js file on disk (possibly as
a result of running `gopherjs build` in the past), it would take higher
precedence than the freshly compiled {base}.js file. Fix that by
rearranging the serving precedence logic as follows:

1. {base}.js && valid Go command in folder -> compiled {base}.js
2. {base}.js.map && valid Go command in folder -> compiled {base}.js.map
3. Look in all Go workspaces (GOPATH and GOROOT)
4. index.html && valid Go command in folder -> our own stub index.html

This way, if there's an index.html file on disk, it takes higher
precedence, but not {base}.js and {base}.js.map files.