Do not include _test.go files when not building tests.
Fixes #260.

Add IsTest bool to PackageData to specify if package is being build
normally or for tests.

parse was made unexported. It was undocumented, and not used by any
external packages. Unless there's a good reason to export it, it's
better not to, as that keeps the API smaller and simpler, allows
internal code to change more easily. For this fix, I changed signature
of parse to accept an extra parameter.