tests: Rename empty.go to doc.go.
An empty .go file is needed because this package is empty aside from
tests. Rename it to doc.go since that's a more common name and it
becomes more readable at a glance (e.g., people can know it will not
contain any code without even having to open/look in it).