tests: Add failing test for internalization/externalization of time.Time/Date.
This is a test for issue #279.

Proper testing for this issue prevents "testing" package from being
imported. So, implement it as a low level single file .go test and
invoke it from backend.

In the time_inexternalization.go, you can use these statements for
manual testing:

	var _ = time.Now()      // Force time.Time not to be DCEed.
	var _ = time.Unix(0, 0) // Force time.Unix (and time.Time implicitly) not to be DCEed.

Adding both of those statements (or just time.Unix one, since it also
includes time.Time indirectly) would cause the test to pass. But a true
fix will be in the compiler itself, in a way that will avoid the
panics/incorrect behavior even when time.Time and/or time.Unix are
DCEed, or time package is not imported at all.