Fix possible panic during internalization/externalization of time.Time/Date.
Fixes #279 (initial report).

The internalization/externalization code only checks if the time
package is present and assumes time.Unix and time.Time are available if
so. Previously, it was possible for time package to be present, but
either time.Unix or time.Time were not (due to dead code elimination),
causing possible panics.

In time package itself, ensure time.Unix func and time.Time struct it
returns are always included (despite DCE) to prevent that from
happening.