compiler: fix runtime exceptions with blocking deferred functions
This issue arises when re-winding the stack to complete a blocking
function call inside a deferred function.  More specifically it occurs
in functions that call blocking functions both with the defer keyword
and in return statements.  A trivial example is given in
[tests/deferblock_test.go](tests/deferblock_test.go),
`ioutil.ReadFile` is an example of such a function from the stdlib.