README: Suggest working with Go packages by default.
When writing Go code that is something more than just a very short
snippet, it helps to organize code as a Go package. It may contain 1 or
more .go files in a directory.

`go build` is primarily used with Go packages, as in `go build
import/path`. We should suggest the same use of `gopherjs build`. That
will make it easier for users to get started and grow their code,
because they'll be able to split into multiple .go files without having
to update their build step.

It's still useful to be able to use build command with ad-hoc packages
specified with a list of .go files, so continue to mention that mode.
Both go and gopherjs tools support it, and it might be more common and
useful with GopherJS.