Clean up.
> But [Content-Length header] doesn't seem to be set even for
non-streaming responses.

That wasn't true. It is set for non-streaming responses. It's just that
CORS needs to allow that header to be accessed by the frontend code, by
setting w.Header().Set("Access-Control-Expose-Headers",
"Content-Length"), etc.

I wish we could use "redirect": "manual", since http.RoundTripper is
not supposed to follow redirects, but that can't be done. Due to
security reasons (/cc @mholt), browsers don't allow frontend code
access to see intermediate redirect URLs, because that might allow for
cross-site scripting attacks (according to
https://fetch.spec.whatwg.org/#atomic-http-redirect-handling).
XHRTransport suffers from the same problem/limitation. So, it's better
to at least succeed and finish the request rather than not being able
to follow redirects at all.

The no body case is not applicable, I added that when I accidentally
confused request body for response body. It should always be set since
we're checking for ReadableStream support before opting in to use
fetchTransport. If that's not the case, we'll need to adjust that logic.