Move comment to a more relevant place.
It does not belong to `return &fetchTransport{}` line, so move it
closer to the relevant line.

Instead of putting it above, put it on the side. This is because this
is an internal detail that you'll only care about if editing or trying
to understand that line better. It's not needed for general readability.

So having to scroll to the right to see it is very fitting IMO.

This is modeled after Bret Victor's style of putting footnotes on the
right hand side of the article (side notes?). They can be skipped
unless the reader wants to learn more. See
http://virtivia.com:27080/5o1yhnwvjrvw.png.