WIP-388: Rewrite gopherjs tool.go to use cmd/go functionality.
DO NOT MERGE. This is a PR for review only, it's not in a mergable state because not all components are present. I want to break review of the final PR into smaller pieces, to make it more diffs smaller and manageable.

This change rewrites tool.go to use the functionality of cmd/go with GopherJS compiler support added, which will be bundled here.

Changes:

- watch flag is removed. Supporting it is possible, but difficult at this time and it's hard to test it. It may be brought back later, or maybe it'll be decided that it's out of scope, and another tool can help with watching. With gopherjs serve command available, the need for -watch flag is greatly reduced.

This is a mostly mechanical change, done by hand, so there's a chance for typos or other screw ups, forgetting flags, etc. This is what I want this review to catch.

Helps #388.