build: Add test to verify that augmented packages don't import extras.
Natives augment the standard library with GopherJS-specific changes.
This test ensures that none of the standard library packages are
modified in a way that adds imports which the original upstream
standard library package does not already import. Doing that can
increase generated output size or cause other unexpected issues (since
the cmd/go tool does not know about these extra imports), so it's best
to avoid it.

It checks all standard library packages. Each package is considered as
a normal package, as a test package, and as an external test package.