compiler/natives/src/net/http: Stay with ignored stream cancel error.
Previously, it was not possible to call stream.cancel and block until a
response was available, because of a bug in GopherJS that prevented
blocking func calls from being deferred.

That bug has since been resolved in 50101461d59f26b13c876f57aab314ce55387671,
so it's possible to do that now.

However, I found no compelling reason to do so at this time. I prefer
not to implement this without a concrete reason and a way of testing
that it works correctly, so simply update the comment and remove
commented out code.

If a need comes up in the future, it can be implemented and tested then.

Do not regenerate because this change is a noop.