build: Allow overriden functions calls
Before this commit, every overriden standard library function was renamed
'_'. Now, if there is the following directive in the function comment:
"//gopherjs:keep_overridden", they are prefixed with "_gopherjs_overridden_"
so that they can be called from the overriding function. This is typically
useful to use the standard library implementation as a fallback.