Handle GOARCH env var similarly to cmd/go.
Externally, GOARCH=js is the only supported mode for gopherjs compiler.
Make it so that setting GOARCH=js explicitly does what one would
expect, which is normal gopherjs behavior instead of a strange error.

If GOARCH is set to any value other than js, which are not supported,
exit with an error (similar to cmd/go behavior).

Set "GOARCH" env var to correct value for gopherjs in tests/run.go.

Fixes #594.
Closes #595.