Use simpler, more readable and consistent control flow.
Execute functionality to get an error, handle the error to get an exit
code, and then exit.

Do that instead of trying to exit with the exit code that is the result
of handling an error of executing a closure.

It's equivalent behavior, but the new version is cleaner IMO.

This is also more consistent with the other occurrences of same pattern.