Use err instead of exit code.
It's equivalent functionality, but simpler and more readable to someone
not familiar with handleError. That person doesn't need to jump inside
handleError and understand how it works in order to understand when
exit code will be 0 or non-zero. It's simpler to just use the err !=
nil value right away (without processing it via a middleware func).