Minor refactor (no behavior change) to simplify and clean up. (#596)
This change is a minor refactor (no behavior change) to simplify and
clean up some of the code I've recently seen/looked over, and found
opportunities to improve. Some of the changes are ones I wanted to make
while working on #595, and some others I just did on the fly when
making this change.

Changes done include:

- Move sprintError under printError.
- Simplify handleError signature to accept error value.
- Use simpler, more readable and consistent control flow.
- Execute closure first, handle error after.
- Use err instead of exit code.
- Replace 'bytes.NewBuffer(nil) -> new(bytes.Buffer)'.
- Remove closure for gopherjs doc command.
- Simplify (some) opportunities found with gosimple:

	tool.go:517:59: should use time.Since instead of time.Now().Sub (S1012)
	third_party/importer/export.go:431:5: should use strings.ContainsAny(format, "{}\n") instead (S1003)
	third_party/importer/import.go:55:2: should replace loop with p.typList = append(p.typList, predeclared...) (S1011)

There should not be any provable behavior changes.