Add local flags to gopherjs run command.
Unlike all other commands, they were not explicitly added to this
command before. It still worked because it used global flags. Now,
it works by using local flags.

This is more explicit and clear than the original version.

It also makes it easier to modify flags on commands in the future.