Make flags local to the commands that actually support them. (#607)
Previously, many of the flags were global, instead of local to each
command. As a result, it was possible to set flags that are unsupported
for a given command.

This change makes all flags local, and only adds flags that are
actually supported by each command.

This improves the output of gopherjs --help and gopherjs <command>
--help, because it more accurately maps to the reality of flags and how
they're used. This way, flags that are unused by a certain command do
not show up in that command's --help output.

It's no longer possible to set -w flag with gopherjs serve, which was
meaningless (it had no effect, and was confusing).

The -w flag is also removed from gopherjs get command. It's unlikely
anyone is using -w flag together with gopherjs get command on purpose,
since it's pretty loaded and confusing. Instead, people can use
gopherjs get, followed by gopherjs install -w to achieve same effect.

Fixes #606.