compiler/natives/src/testing: Use a basic but working callerName.
The testing.callerName function in Go 1.9 tries to use runtime.Callers
to figure out the calling function name (qualified with a package path).
Since the current runtime.Callers is not implemented for GOARCH=js, and
returns 0, it causes testing.callerName to panic.

Replace testing.callerName with a basic implementation that allows
tests to run. A TODO comment indicates that this should be revisited
and improved, if possible, to print better output.

Fixes:

	Error: testing: zero callers found