WIP: nosync: Implement sync.Map.
This is a new type in Go 1.9.

Map is a concurrent map. It's supposed to be safe for multiple
goroutines to call a Map's methods concurrently.

For now, implement it as a simple wrapper around map[interface{}]interface{}.
Since GopherJS is not multithreaded, as long as context switches don't
happen within Map methods, this should be safe.

TODO: Verify this is actually safe/correct.

Fixes:

	$ gopherjs test --short reflect
	$GOROOT/src/encoding/json/encode.go:335:18: Map not declared by package nosync