CI: Prefer --minify flag for all tests.
This largely reverts commit b9c0354ebade7f285afbedee31abd6f8dbe9b3c3.

It was originally done for convenience, to improve readability of stack
traces during failures.

However, it's more important to prioritize doing test coverage of the
same build mode as used in production (minification on). That's a
sufficient reason to keep the --minify flag for gopherjs test of all
packages.