compiler: Replace old importer with x/tools/go/gcimporter15.
The copied importer is an old package from Go 1.5 times. It was a copy
of x/tools/go/importer package. It does not support type aliases added
in Go 1.9. Encountering a type alias declaration caused a panic.

Use the gcimporter15 package instead. It supports a more modern binary
export format and includes support for type aliases.

The gcimporter15 package is marked as deprecated and said to be deleted
in October 2017. However, the replacement package does not expose the
functionality suitable for GopherJS current needs. That means we'll
need to find a resolution by the time it's deleted. It will either be a
feature request to add the needed functionality, a rewrite of GopherJS
code to use available functionality, or vendoring/copying the code.

Updates #278.