build: unify setting of ulimit and stack_size
Picking up on @shurcooL's comment in
https://github.com/gopherjs/gopherjs/pull/669#issuecomment-325138883.

We've started randomly seeing std library tests tip over the V8 stack
size limit, causing text/template TestMaxExecDepth to fail randomly
locally and on CI. This had previously been addressed by @neelance in
1f89545; the --stack_size flag was passed to NodeJS which in turn passed
the value onto V8. But per
https://github.com/nodejs/node/issues/14567#issuecomment-319367412 it
was pointed out that the value of ulimit -s must be >= the value of
--stack_size for the --stack_size to make any sort of sense. Hence this
commit also harmonises the setting of ulimit -s during the CI test run
with the value of --stack_size that is passed to NodeJS (which in turn
passes that value onto V8) when running either gopherjs test or gopherjs
run.