js: Make test cases non-fatal.
There's no reason for these to be fatal. We can continue running the
test further if any one of them fails. That way, if there are multiple
failures, it's easier to see them all at once.

Separate var x from the switch by a blank line, since it's used by both
switch and the if statement below. Having it without a blank line makes
it seem that it's only needed for the switch.