Compute binary name from base of directory rather than import path. (#702)
When building a package that lies outside of a GOPATH, its import path
will be local, and its base might be unsuitable for figuring out the
binary name. Instead, use the base of the absolute path to the
directory where the package is located.

Do so even for packages that are inside a GOPATH workspace. It should
produce identical results. As of right now, there are no known cases
where that's not true, but if any are discovered and reported, we can
fall back to using base of import path as long as
build.IsLocalImport(pkg.ImportPath) returns false.

Fixes #699.