syscall_unix: fallback to process.exit when available
"go test" currently always fails with "fatal error: all goroutines are
asleep - deadlock!". In fact, all programs that use "os.Exit" will fail
with this.

It turns out that the current "runtime.Goexit" function throws an
exception which propagates up which violates the expected behavior of
the exit syscall (which should not return). Therefore call
"process.exit" when available.

Fixes #626