build: Add "purego" build tag to default context.
The proposal in golang/go#23172 was accepted. The "purego" build tag
is intended to be a community agreed upon soft-signal to indicate the
forbidden use of unsafe, assembly, or cgo.

Since GopherJS doesn't support those, it should always use "purego"
versions of Go packages when they're available. This is done by always
setting the "purego" build constraint.

Fixes #746.