build: Add "purego" build tag to default context. (#747)
The proposal in golang/go#23172 was accepted. The "purego" build tag
is intended to be a community agreed upon soft-signal to indicate the
forbidden use of unsafe, assembly, or cgo.

Since GopherJS doesn't support those, it should always use "purego"
versions of Go packages when they're available. This is done by always
setting the "purego" build constraint.

Document build tags that are set.

Bump GopherJS version to 1.9-2. Setting the "purego" build constraint
can change behavior in theory. It's better to have a distinct GopherJS
version to attribute it to.

Fixes #746.