compiler/natives/src/encoding/gob: Skip TestTypeRace.
TestTypeRace was added in Go 1.10, and it needs to be skipped because
it relies on sync.WaitGroup to behave correctly under concurrency.
However, encoding/gob is currently configured to use nosync, which
cannot succeed.

Since this is a new test, opt to skip it for now rather than making it
succeed by replacing nosync with sync.

Related to golang/go@1a9f27d5034056c1adf53a39939973c4495370f4.

Fixes:

	=== RUN   TestTypeRace
	--- FAIL: TestTypeRace (0.00s)
		Error: sync: WaitGroup counter not zero
	FAIL    encoding/gob    3.030s