compiler/prelude: Initialize high/max before checking range
In $subslice function in prelude, it was checked whether the arguments high/max were in the given slice's range. This was done even when high/max were not given, and the range check didn't work in this case. This commit fixes this issue by ensuring high/max values to use the slice's length and capacity values as default values.

Fixes #732.