build: Make fewer copies of NewBuildContext.
Call NewBuildContext at the top of Import, ImportDir, and inside
NewSession.

Previously, it was called more often than necessary (instead of being
preserved for entire build session).

Pass build context by value to importWithSrcDir so it can be safely
modified there, without the changes persisisting from one package
to the next.

The build context will need to be available in more places in an
upcoming commit. Storing it in Session helps with that.