js: add MakeFullWrapper to expose exported methods and struct fields.
Currently the documentation for js.MakeWrapper is:

> "MakeWrapper creates a JavaScript object which has wrappers for the
exported methods of i. Use explicit getter and setter methods to expose
struct fields to JavaScript."

Where the value a struct value (or more
interestingly a pointer to a struct value) we can actually auto-generate
getters and setters for exported fields in the JavaScript world, rather
than requiring explicit getters and setters to be defined on the Go
side.

We do this via a new MakeFullWrapper method.