ci: parallelise the running of gopherjs tests
Currently we run the gopherjs tests in sequence; this is limited to a
single CPU core because we ultimately run the tests via Node. We can
improve on this by concurrently running gopherjs test per package.

Also implemented a fix discussed in a previous issue/PR to make our list
of package exclusions explicit, via std_test_pkg_exclusions