natives/internal/cpu: Define CacheLineSize constant for GOARCH=js.
In Go 1.11, internal/cpu package is being used in more places
(including runtime), so it's no longer viable to avoid importing it.
Instead, fix the build by defining the CacheLineSize constant for js
architecture.

Set its value to 0 because it's not helpful in the context of GopherJS
to set it to a larger value. Internal field alignment isn't a factor.

Fixes:

	$ gopherjs build runtime
	internal/cpu/cpu.go:20:16: undeclared name: CacheLineSize
	internal/cpu/cpu.go:37:16: undeclared name: CacheLineSize
	internal/cpu/cpu.go:50:14: undeclared name: CacheLineSize
	internal/cpu/cpu.go:62:14: undeclared name: CacheLineSize
	internal/cpu/cpu.go:70:15: undeclared name: CacheLineSize
	internal/cpu/cpu.go:95:15: undeclared name: CacheLineSize
	internal/cpu/cpu.go:101:19: undeclared name: CacheLineSize
	internal/cpu/cpu.go:118:19: undeclared name: CacheLineSize