CI: Upgrade to CircleCI 2.0. (#864)
This is a minimal diff change to upgrade from CircleCI 1.0 to 2.0
and get it working, since 1.0 has been EOLed. The goal was to keep
the config.yml file as simple as possible for now; we can consider
changes and enhancements to it as part of future work.

Remove os/user package from being tested, since it's failing in the
CircleCI 2.0 environment (see tracking issue #863).

Resolves #766.
Updates #797.