Add syscall/js (#908)
This PR adds a new package `syscall/js`. Originally `syscall/js` is for WebAssembly, but in GopherJS
this package works as a wrapper of `github.com/gopherjs/gopherjs/js`. With this package,
developers can use `syscall/js` both for GopherJS and WebAssembly.

The implementation is basically same as `github.com/gohperjs/gopherwasm`.

Fixes #899