build, compiler/natives/src/reflect: remove Go 1.11-specific code
GopherJS 1.12 is out by now, so it's safe to remove this code that
was added to deal with differences between Go 1.11.0 and Go 1.11.1.

There should not be any change in behavior, except the internal and
temporary go1.11.1 build tag will no longer be set by GopherJS.
It was never intended to be used externally, so hopefully nobody did.

Regenerate ./compiler/natives.

Updates #862.