Start a v2 of gopherjs/build package.
For the time being it will coexist with the current implementation and
will be activated by setting an environment variable:
`GOPHERJS_BUILD=go/packages`, similar to how experimental features are
activated in Go. Eventually it will replace v1 version of the package
entirely.

Example of using build command with the new package:

     $ gopherjs build --sourcemap=false --color=false errors
     failed to build []: building is not implemented yet