Fix up node-syscall to work with Node 12.
Some APIs were removed in Node 12, and new "safer" APIs must be used
instead. Perhaps this implementation is not the most elegant or produces
less then ideal error messages, but at least it compiles.

While at it, added an arena type to manage and gracefully free temporary
buffers we allocate for the duration of the call.

I couldn't find any tests for the syscall module, so I wrote a small
one.