fixes blocking when using function literals
When a function literal is used that calls another function and
that other function has yet to be analyzed, the call site for
the function literal is not retained, so it's not processed later,
so the function that was calling the function literal is not made
blocking itself, when it should be.

A new field is used in info.go. It's name kept short to keep gofmt
from readjusting surrounding lines. It's not the most descriptive name
as it stands.

A short test case that demonstrates the problem can be found in
issue #955.

Fixes #955.