Support compiling Go 1.16 `runtime` with GopherJS. (#990)
Support compiling Go 1.16 `runtime` with GopherJS.

The vast majority of the code (even in errors.go) is not used by
GopherJS, so starting with this commit the package is entirely defined
by our own implementation in natives. As a bonus, this allowed to drop
a few unexported symbols that existed purely to let other unused
functions to compile.