Fix `time` package to compile with Go 1.16.
I'm not sure of the role of the seq field plays outside of the time
package, but I'll add it there. In the upstream code the runtimeTimer
struct seems to replicate (and require to be compatible with)
runtime.timer struct, however in GopherJS it is clearly not and it
doesn't seem to be a problem.

Similarly, in the upstream startTimer, stopTimer, resetTimer and
modTimer are actually implemented in the `runtime` package, but since we
already re-implemented two of them here, I'll just keep this package
self-contained.