Add `internal/reflectlite` package support.
This package is meant to be a cut-down version of the `reflect` package,
which is used by packages like `errors` to avoid a cyclic dependency
with `reflect`.

This code is based on goplusjs/gopherjs@8e608e5d486b4e9f5142c6c6822649856f936421.

The natives code of the reflectlite package is a subset of its
counterpart from reflect.  Unfortunately, this requires us to duplicate
a substantial chunk of code with all the maintenance problems that stem
from it. I have considered two alternative approaches, none of which
worked out:

  - Automatically rewrite imports from `internal/reflectlite` to
    `reflect`. Unfortunately, this creates cyclic dependencies and
    causes compilation to fail.
  - Redefine types in reflectlite as aliases of their reflect
    counterparts. This both causes a cyclic dependency and creates a
    divergence between GopherJS and Go standard library. TinyGo seems to
    be doing that somehow, but I suspect that they are able to handle
    cyclic dependencies.

As a bottom line, duplicating a subset of reflect code is the lesser
evil that is available to us. In the long term we could try and converge
our reflect implementation with the upstream, so that this is no longer
our problem.

Co-authored-by: visualfc <visualfc@gmail.com>