Implement `runtime.FuncForPC()` for GopherJS.
`FuncForPC()` returns function/file/line metadata for any given code
position (effectively a memory address). However, since GopherJS runtime
doesn't have access to raw addresses, we emulate this by assigning
virtual position counters for call places seen by Caller() and use them
to retrieve the original function information.

This is a more limited implementation that the upstream, but I expect it
to work for most of the real-life cases, and it fixes a few test
failures in the `reflect` package. With this it should be fairly easy to
bring back proper support for `testing.T.Helper()` function.

Co-authored-by: Jonathan Hall <flimzy@flimzy.com>