Provide a working implementation of `reflect.StructOf()` function.
It has been introduced all the way back in Go 1.10, but GopherJS never
properly supported it. The new implementation seems to pass most of the
tests, which is an improvement. There are a couple of tests that are
still failing, but fixing that seems like a harder task.

The code is based on goplusjs/gopherjs@a293d52579eda792a74b30b945fee0691b91372d.

Co-authored-by: visualfc <visualfc@gmail.com>