Only `throw null` in $callDeferred() when called directly from $panic()
$callDeferred() uses `throw null` to interrupt normal execution when
called from a panic(). However, it simply returns when called orderly
at the end of a function.

A tricky interaction may happen when one of the deferred function blocks
on a channel or mutex. The goroutine will get paused and when it
unblocks, deferrals will continue to execute via $callDeferred() at the
end of the function. Under that circumstance `throw null` trick is
unnecessary since the main body of the function is outside of the call
stack anyway.

This fix allows fixedbugs/issue40629.go test to pass successfully.