Use build.Default.GOARCH for syscall package.
So long as we rely on OS-specific syscalls, we must build the package
with OS-appropriate GOOS/GOARCH (see also
https://github.com/gopherjs/gopherjs/issues/693). Using runtime.GOOS
here is not ideal, since it ignores environment variables, making
"cross-compiling" impossible.

This is not a big deal in most cases, but makes reproducing errors like
in https://github.com/gopherjs/gopherjs/issues/1027 more difficult.