Add a compiler directive to control function body pruning.
Fixed https://github.com/gopherjs/gopherjs/issues/1033, which was caused
by incorrect import pruning, which was introduced in
https://github.com/gopherjs/gopherjs/commit/58c1058cb3db0d3a6d0303037051f1db057e29a9.

The original intent was to prune all function bodies, which were
replaced by the corresponding overlay. That, in turn, required pruning
imports which might have become unused in the modified AST tree.
Unfortunately, checking whether the import is used requires knowing the
imported package name, which may not match the last component of the
import path and is not readily available at package augmentation stage.

This change drops import pruning and introduces the
`gopherjs:prune-original` directive to only trigger function body
pruning selectively to prevent compile errors when it doesn't lead to
unused imports. When it does, a different solution is required, like on
the `TestGarbageCollection` case.