Handle circular references when internalizing JS objects.
Fixes https://github.com/gopherjs/gopherjs/issues/968.

We keep track of values we've already internalized (accounting for the
type to which it was internalized) and if the value we've already
encountered is in the map, we return its internalized value from the
map.

This means that the returned object will also contain a circular
reference.